Might as well improve challenge sector icons too. (doesn't work rn)

+++ src/fos/maps/bases/UxerdBaseGenerator.java
@@ -1,11 +1,10 @@
-package fos.maps.generators.bases;
+package fos.maps.bases;
 
 import arc.math.*;
 import arc.math.geom.*;
 import arc.struct.Seq;
 import fos.world.blocks.environment.UndergroundOreBlock;
-import mindustry.ai.*;
-import mindustry.content.Blocks;
+import mindustry.ai.BaseRegistry;
 import mindustry.game.*;
 import mindustry.gen.Building;
 import mindustry.type.*;
@@ -13,19 +12,17 @@ import mindustry.world.*;
 import mindustry.world.blocks.defense.Wall;
 import mindustry.world.blocks.payloads.*;
 
-import static mindustry.Vars.*;
+import static mindustry.Vars.content;
 
-public class LumoniBaseGenerator extends FOSBaseGenerator {
-    public LumoniBaseGenerator() {
+public class UxerdBaseGenerator extends FOSBaseGenerator {
+    public UxerdBaseGenerator() {
         schematics = new String[]{
-            //cores
-            "bXNjaAF4nF2QsW7DMAxET7YjVLCj1EPRr9CQsd9SdHBipTWg2IGUNL9fnrgVAvxoiiceiRFji26drhHv6XFd1imUnynPcQ7nLcdwDEf0cyznvNzuy7YCsGk6xVTQfH7t8HbZSihL+o05PKeUQpryd8ThXxqvTDyXdQ73Rz4ta4Rj5vZIJWJkWNtdtnzPsRRp8wEYOS3YUuEIA4/WoJHYEJ1iB9MILEA4NEz2Wjko9kQrgXw6ORUDahNPQcc7vmgVTpP1bsc7wjOwtMNuYoPoFD1NWGlkWHKA5F/kVOtV4FTgVOBU4FTgKDCSkgFa8dWhTmAVjnYHtdurpYGVRP3b849DVsGeAiarXc8Z6+K4LK8789wZ0SsGRfXia60Ysgp57A+okDUt",
+            //core
+            "bXNjaAF4nGWS7XKaUBCGF1BE/ASVINFMb4AfvZ5OfyCeJE6NZA44yeX0UrvveU8nnWnM8Li7736KPMm3SEa35s1Idv809lyfmt7UbWdN/V1mZ9O39vI+XLqbiMTX5mSuvYQ/fqayfu762nbtr7q19/7VWHr696Y19clezi9GcnheTTPUL+ZmbDN0VubwDZdb/XGxRpZfOed7O8gD7PZ+MvZyf6s/muu1vjZWS2X/BaRwE5i+u1tNN5+DbVp0+GcO290HnYzJ2Om5s4Om9LrNb/n6i0QC/YSi//qQAIjgjyRMJALHCZQjPAJFSMuJRi6oT3hV6bUxJTG1MbWxCwYyoXbizFASdk0oSpw3kqnvPnX2WJlIECtSYkbMJUBsgTYpB0zZNGW91A84Y9OZM2NlIuFE02eBy4JnCd2curnXzaEDXHDB4MLPuJBYhwInnonn1A27QBaworWWYKSukIiIETEmYmJCJMRUgkSR0poRuvlUCy9xgCVKx2olxBSSFQ4CK6dzIxGwlWik8hT7rFEFWNOZSZQoclobwifsiIJ4IEpiT1TEI3Egjmib8dz533Pja+a+uuAWw2RaLQSOuHjOw+eIiaJALIcE1p54pPNI6wkJG+RFioV7fTay9Fx5bjy3njvVB8qS0Fpj7ekOucWPBWuDX3XrpKGyQIetlFoB5p7eR3qP3uum2eE1BGbEnFgytqaVETnm2bkmkbYARVlgroKLF1y8QC84j7SekPHgMkbKkjjCW/o6JQ9Ysk7JOiUPWLJOyTp7f5E9L7JnncovX3H5yi9fcfmKy1d++QqVRF8AN/uBPQ/seeDsB/Y8QPkH4rpr9Q==",
 
-            //defenses
-            "bXNjaAF4nF1QzU7GIBDc/tAWqtGLifEdOHjxZYwHbPn8SGj7Bai+vrtMvJgWhl2W2dkhQ3NH/e42T8/x3MLubL66tPrVrv7i9+ztK82rz0sKtxKOnYiG6D59zNS+fwz0cDmyzSF++2R/XIz0KImbSyUsZ3TlSPT0r8RGl7483Um6hN0yWaH7v2hNgVnmUPxm83GmxXPLN17UUCfQAjpAD1CAATACJoAGGIGGv7apHG3LZ34n0cSLIwN+yfBv0IYj6aWoERg4ZjDo3QhLL3Tsohxkq5U9iyC5qyJ6PFAgU1IpMCJpIL0qE+mttFHCOchEkqwCB1SOaDti6BHJCaoniRSPzBu9dDJ8zWvYpmGbhm0aDBq2adimoVjDNi22/QIzGTiq",
-            "bXNjaAF4nDWQXW6DMBCEFwNOHEzz26pSz+CXnqD3qPpAwFEtERMZo16/u4zywuddzw7roZbeSqpid/f0Pi73EDs3/3Zp8IMb/M3H2btPagY/9yk8cpgiEemxu/pxJvX9o+lwm2b36FIO/TJ2eUpkpZNDdCzK1D6rIYVxhHxKnt2z70W+dv5CHFxe0jVET03I/u7maUm95999ERWkqShIcbWiABRQAhXf8tlIVfJRCRRQoVkDGtiwD9+0VGJ4BdsVjErmaoHIt6KraSd3NR1loZpOqM7ABXgVFw0zDTMtLhWjRmWBFngB9sABOAJnWUKzJ2Oz7sabrHMblijFOAEXWWkrAQgUsL5hK8qCU1mjM4jOIDqD6AyUBp4GngaeO4w3T5yl2fBK/BorH/rgvsW1RSYWmVhkYiWTf2phOYk=",
-            "bXNjaAF4nEWPW07DMBBFr2Pn1YSSdgVswB+sB/HRNkZYcpzKcVq2wzd7Yw3MZEBoZJ95eeYaOxw1TDxNDvW7C/7DPqMb3XJJ/pr9HAFU4XR2YUH58v359VpjeJsXe/dxtHlNZx8des5kHymbHFqOtlk4suuna5pvbrRj8iH8N9PQjP1flOY1u4TOZzfZZV7TxdHuJzowgCIrGqEmFsRf9IIB0GSK2zeHYQSloBLUggaqJLTQHO0k2Ql6qT2g4GgvyUfBILUDtsGGtxvRYkSLYS2KlzasvuJLEUR/vekHvdBca3Eo+Zs/qmA1Qw==",
-            "bXNjaAF4nE3RXU6EMBQF4Ftob8vPOMZ98OAe3IXxAYaakIxAADNxOT67N9cg5ZxJJClfofe2JyBBzrnYsf2I4vppa56l6uN6WYZ5G6ZRRPTadvG6inv9/f55s+Lfp7XZK+WcJpdpnuPS3IYlymN6cRvGvtk+l24Yozz9K+nabYvL177ji+Ay+8jCPnZzPLpwTEWBBwGLBWtLPNYB25xSkUGLQYtBSyZ3Dc1oTi11VKmngRa0pBWt6Yk+HJlS5nRImqSbAg8C1kpUHuktSixKLFI7uWtoRnNqqaNKPQ20oCWtaE1PFOkV6RXpFdEU0RTpFek1pU+fHME8f59Hi0eLFy4WNJ36ByHMKHA=",
+            //resources
+            "bXNjaAF4nE1Oy07DMBCcOo+mpJSKAxyQ+AMf+B7EwXG2rUUaV2tHlL9nF1cCWfZ4ZmdnF3u0FerZnQkvy5V4tINLZGkawnK2F46eUrJv6EdKnsMlhzgDaCc30JRg3j822B9isuL8tJ6XdCJGHzKdbYoLeyrldHGe7MBhPBJ2f8q4+IxH5Sdy2R5pJnY5Mraq5TDbr8CEp98RVBItXTM7r65/2RyXLKMfVPGRSZ4pzt+y7KtcGGAl0DTAsxD9oxJYyatFoBYwxVoXLqZGoOrUWqtqRNAUOUXVukBb4lqNMwIGVSOgHqWSUWFdOte3zq7Q7kY3hW5u9E5pix5GN11JV19GbAvcK3TY6bo/ajhNMQ==",
+            "bXNjaAF4nCXOzW6DMAwH8D8E6AcFSg/TpD3BDjnseaYd0uCuaECqBKbt7WfPkcgP29gGPS4GxeJmwvP2Q3GwV5fIpnH6pmjncSH7hnqg5OP4WMewAKgmd6UpIX//2OHpFpINkexAK/k1RJtmN02ox5Vmm8IWPeEsH8Xgv6yPW7pTxEUyd3Kr/aSFouNGtJJLD+d52OZXNBKv42KHOPLETkIvq3yYwvLLf/LKD07IM6YvgRfkyNAiN/zC+QKGo0zolLPSg68CDUzJdDAFw8kCpVCi4nYjtFKrtLZT9rJVyKS251X/GKVQSqVSdspeOShHpVZOSqO0SqeclV446PajUisnpVFaoeJmPn/KJUGC",
         };
         init();
     }
@@ -42,7 +39,7 @@ public class LumoniBaseGenerator extends FOSBaseGenerator {
 
         float bracketRange = 0.17f;
         float baseChance = Mathf.lerp(0.7f, 2.1f, difficulty);
-        int wallAngle = 180; //180 for full coverage
+        int wallAngle = 70; //180 for full coverage
         double resourceChance = 0.5 * baseChance;
         double nonResourceChance = 0.0005 * baseChance;
         BaseRegistry.BasePart coreschem = coreParts.getFrac(difficulty);
@@ -50,9 +47,9 @@ public class LumoniBaseGenerator extends FOSBaseGenerator {
 
         Block wall = getDifficultyWall(1, difficulty), wallLarge = getDifficultyWall(2, difficulty);
 
-        for(Tile tile : cores){
+        for(Tile tile : cores) {
             tile.clearOverlay();
-            Schematics.placeLoadout(coreschem.schematic, tile.x, tile.y, team, false);
+            Schematics.placeLoadout(coreschem.schematic, tile.x, tile.y, team, true);
 
             //fill core with every type of item (even non-material)
             Building entity = tile.build;
@@ -61,21 +58,21 @@ public class LumoniBaseGenerator extends FOSBaseGenerator {
             }
         }
 
-        for(int i = 0; i < passes; i++){
+        for(int i = 0; i < passes; i++) {
             //random schematics
-            pass(tile -> {
-                if(!tile.block().alwaysReplace) return;
+            pass(800, tile -> {
+                if (!tile.block().alwaysReplace || Mathf.within(tile.x, tile.y, 350, 350, 100)) return;
 
-                if(((tile.overlay().asFloor().itemDrop != null || tile.overlay() instanceof UndergroundOreBlock || (tile.drop() != null && Mathf.chance(nonResourceChance)))
-                    || (tile.floor().liquidDrop != null && Mathf.chance(nonResourceChance * 2))) && Mathf.chance(resourceChance)){
+                if (((tile.overlay().asFloor().itemDrop != null || tile.overlay() instanceof UndergroundOreBlock || (tile.drop() != null && Mathf.chance(nonResourceChance)))
+                    || (tile.floor().liquidDrop != null && Mathf.chance(nonResourceChance * 2))) && Mathf.chance(resourceChance)) {
                     Seq<BaseRegistry.BasePart> parts = forResource(
                         tile.overlay() instanceof UndergroundOreBlock uo ? uo.drop :
                             tile.drop() != null ? tile.drop() : tile.floor().liquidDrop
                     );
-                    if(!parts.isEmpty()){
+                    if (!parts.isEmpty()){
                         tryPlace(parts.getFrac(difficulty + Mathf.range(bracketRange)), tile.x, tile.y, team);
                     }
-                }else if(Mathf.chance(nonResourceChance)){
+                } else if (Mathf.chance(nonResourceChance)) {
                     tryPlace(parts.getFrac(difficulty + Mathf.range(bracketRange)), tile.x, tile.y, team);
                 }
             });
@@ -83,7 +80,6 @@ public class LumoniBaseGenerator extends FOSBaseGenerator {
 
         //small walls
         pass(tile -> {
-
             if(tile.block().alwaysReplace){
                 boolean any = false;
 
@@ -132,24 +128,6 @@ public class LumoniBaseGenerator extends FOSBaseGenerator {
                 curr.setBlock(wallLarge, team);
             }
         });
-
-        //clear path for ground units
-        for(Tile tile : cores){
-            Astar.pathfind(tile, spawn, t -> t.team() == state.rules.waveTeam && !t.within(tile, 25f * 8) ? 100000 : t.floor().hasSurface() ? 1 : 10, t -> !t.block().isStatic()).each(t -> {
-                if(!t.within(tile, 25f * 8)){
-                    if(t.team() == state.rules.waveTeam){
-                        t.setBlock(Blocks.air);
-                    }
-
-                    for(Point2 p : Geometry.d8){
-                        Tile other = t.nearby(p);
-                        if(other != null && other.team() == state.rules.waveTeam){
-                            other.setBlock(Blocks.air);
-                        }
-                    }
-                }
-            });
-        }
 
     }
 }
